Pequeno refatoramento

+++ frontend/src/pages/profile/index.tsx
@@ -24,7 +24,7 @@ const Profile = () => {
 
     const history = useHistory();
     const ongId = localStorage.getItem('ongId')
-    const ongName = localStorage.getItem('name')
+    const ongName = localStorage.getItem('ongName')
 
     useEffect(() => {
         api.get('/profile', {
@@ -41,13 +41,9 @@ const Profile = () => {
         history.push('/')
     }
 
-    async function handleDeleteIncident(id: any) {
+    async function handleDeleteIncident(id: number) {
         try {
-            await api.delete(`/incidents/${id}`, {
-                headers: {
-                    Authorization: ongId,
-                }
-            })
+            await api.delete(`/incidents/${id}`, { headers: { Authorization: ongId }})
             setAux(aux+1);
         } catch(err) {
             alert('Erro ao deletar caso, tente novamente')
@@ -57,34 +53,34 @@ const Profile = () => {
     return (
         <div className="profileContainer">
             <header>
-                <img src={ logoImg } alt="Be the Hero"/>
-                <span>Bem vinda, {ongName} </span>
+                    <img src={ logoImg } alt="Be the Hero"/>
+                    <span>Bem vinda, { ongName } </span>
 
-                <Link className="button" to="/incidents/new">Cadastrar novo caso</Link>
-                <button type="button" onClick={handleLogout}>
-                    <FiPower size={18} color ="#E02041" />
-                </button>
+                    <Link className="button" to="/incidents/new">Cadastrar novo caso</Link>
+                    <button type="button" onClick={handleLogout}>
+                        <FiPower size={18} color ="#E02041" />
+                    </button>
             </header>
 
             <h1>Casos Cadastrados</h1>
             <ul>
                 {
                     incidents.map((incident) =>                       
-                        (<li key= {incident.id}>
-                            <strong>CASO:</strong>
+                            (<li key= {incident.id}>
+                                <strong>CASO:</strong>
                                 <p>{incident.title}</p>
 
-                            <strong>DESCRIÇÃO</strong>
-                            <p>{incident.description}</p>
+                                <strong>DESCRIÇÃO</strong>
+                                <p>{incident.description}</p>
 
-                            <strong>Valor</strong>
-                            <p>{Intl.NumberFormat('pt-BR', { style: 'currency', currency: 'BRL' }).format(incident.value)}</p>
+                                <strong>Valor</strong>
+                                <p>{Intl.NumberFormat('pt-BR', { style: 'currency', currency: 'BRL' }).format(incident.value)}</p>
 
-                            <button onClick={() => handleDeleteIncident(incident.id)} type='button'>
-                                <FiTrash2 size={20}/>
-                            </button>
-                            </li>)
-                        )
+                                <button onClick={() => handleDeleteIncident(incident.id)} type='button'>
+                                    <FiTrash2 size={20}/>
+                                </button>
+                                </li>)
+                            )
                 }
             </ul>
         </div>
